[refactor/#194] 메시지 파일 분리
