+++ src/hooks/auth/useUpdateUser.ts
@@ -1,5 +1,5 @@
 import authAxios from '@/lib/api/authAxios'
-import { SAVE_ERROR_MESSAGE } from '@/lib/data/constants'
+import { SAVE_ERROR_MESSAGE } from '@/lib/data/messages'
 import { useUserStore } from '@/lib/stores/userStore'
 import { UpdateUserValues, User } from '@/lib/types/userTypes'
 import handleError from '@/lib/utils/errorHandler'
