+++ src/lib/data/messages.ts
@@ -1,0 +1,15 @@
+export const DEFAULT_ERROR_MESSAGE = {
+  title: '오류',
+}
+export const SIGN_IN_ERROR_MESSAGE = {
+  title: '로그인 실패',
+}
+export const AUTH_USER_ERROR_MESSAGE = {
+  title: '사용자 로드 실패',
+}
+export const SIGN_UP_ERROR_MESSAGE = {
+  title: '회원 가입 실패',
+}
+export const SAVE_ERROR_MESSAGE = {
+  title: '저장 실패',
+}
